Create ChangeableAlias.js
Adding the ChangeableAlias, an alias that can have its wrapper property set at any time. As a result it does not have to be initialised with another property when it is created. This can be useful if you have a property which you cannot create in the constructor, but which you still need to exist to bind.